Remove VRE v2 form_id and all references for legacy model
