--- spec/fixtures/pdf_fill/28-1900-V2/kitchen_sink.json
@@ -1,36 +1,0 @@
-{
-  "yearsOfEducation": "12",
-  "isMoving": true,
-  "newAddress": {
-    "country": "USA",
-    "street": "1000 5th Ave",
-    "street2": "Apt 812",
-    "city": "New York",
-    "state": "NY",
-    "postalCode": "10028"
-  },
-  "veteranAddress": {
-    "country": "USA",
-    "street": "4201 Henry Ave",
-    "street2": "8th Floor",
-    "city": "Philadelphia",
-    "state": "PA",
-    "postalCode": "19144"
-  },
-  "mainPhone": "2159512800",
-  "cellPhone": "2159516800",
-  "email": "sam@test.com",
-  "veteranInformation": {
-    "fullName": {
-      "first": "Adam",
-      "middle": "B",
-      "last": "Huberws",
-      "suffix": "Jr."
-    },
-    "ssn": "666-33-7777",
-    "VAFileNumber": "88246556",
-    "dob": "1966-05-16"
-  },
-  "privacyAgreementAccepted": true,
-  "signatureDate": "2025-06-26"
-}
