--- spec/fixtures/pdf_fill/28-1900-V2/overflow.json
@@ -1,38 +1,0 @@
-{
-  "yearsOfEducation": "12",
-  "isMoving": true,
-  "newAddress": {
-    "country": "USA",
-    "street": "100117 5th Ave",
-    "street2": "Apt 812",
-    "street3": "6th Floor",
-    "city": "New York",
-    "state": "NY",
-    "postalCode": "10028"
-  },
-  "veteranAddress": {
-    "country": "USA",
-    "street": "4201 Henry Ave",
-    "street2": "8th Floor",
-    "street3": "Suite 106",
-    "city": "Philadelphia",
-    "state": "PA",
-    "postalCode": "19144"
-  },
-  "mainPhone": "2159512800",
-  "cellPhone": "2159516800",
-  "email": "sam@test.com",
-  "veteranInformation": {
-    "fullName": {
-      "first": "AdamAdamAdamAdam",
-      "middle": "B",
-      "last": "Huberws",
-      "suffix": "Jr."
-    },
-    "ssn": "796-10-4437",
-    "VAFileNumber": "88246556",
-    "dob": "1950-10-04"
-  },
-  "privacyAgreementAccepted": false,
-  "signatureDate": "2025-06-25"
-}
