--- spec/fixtures/pdf_fill/28-1900-V2/simple.json
@@ -1,15 +1,0 @@
-{
-  "yearsOfEducation": "10",
-  "isMoving": false,
-  "email": "sam@test.com",
-  "veteranInformation": {
-    "fullName": {
-      "first": "Adam",
-      "last": "Huberws"
-    },
-    "ssn": "666-33-7777",
-    "dob": "1966-05-16"
-  },
-  "privacyAgreementAccepted": true,
-  "signatureDate": "2025-07-02"
-}
